Make emplace conforming with std containers
Make emplace method of dense_hash_set and dense_hash_map conforming with
the same method of std containers - it takes the same arguments as
value_type constructor. Because in hashing containers we first need a
key before we could construct a value, for dense_hash_map it is not as
straightforward as for simpler containers, but it could be done
relatively efficiently (relying on move semantic of key type).